+++ ccbi-330/configs/mcm-config.json
@@ -3,6 +3,8 @@
     "count_path": "data/mcm_mcm_param.mol_counts.summary.tsv",
     "output_prefix": "test/mcm-binary",
     "cancer_type": "crc",
+    "bad_cohorts": ["L2AV-F", "L2AV"],
+    "bad_batches": ["LTO_232"],
     "total_iterations": 5,
     "iteration_start_seed": 0,
     "binary": true,
@@ -10,5 +12,6 @@
     "do_transform": true,
     "scaler_str": "preprocessing.RobustScaler()",
     "tumor_normal_ratio_min": 1.5,
+    "num_digits": 4,
     "regressor_str": "linear_model.LogisticRegression()"
 }
